requirements.txt
python-telegram-bot==20.5
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
finnhub-python==1.4.4
python-dotenv==1.0.0
APScheduler==3.10.4
aiohttp==3.9.1
SQLAlchemy==2.0.23
pytz==2023.3
